+++ com/vinod/Main.java
@@ -1,10 +1,10 @@
-package com.company;
+package com.vinod;
 
 public class Main
 {
     public static void main(String[] args)
     {
-        String hello = "Hello world";
+        String hello = "Hello Vinod";
 	System.out.println("hello world");
         System.out.println(100);
 	System.out.print("hello world "+hello);
