+++ com/vinod/math.java
@@ -1,4 +1,4 @@
-package com.company;
+package com.vinod;
 
 public class math
 {
