+++ com/vinod/random until code
